requirements.txt
Flask==3.1.0
google-cloud-firestore==2.20.1
cryptography==44.0.1
itsdangerous==2.2.0
requests==2.32.3